Added in couts for easier debugging

+++ Sandbox/Source/Systems/LifespanSystem.cpp
@@ -36,7 +36,8 @@ namespace DeltaEngine
     });
 
     for (auto& ref : DestroyedEntities)
-      em.DestroyEntity(ref);
+	  if(em.IsEntityValid(ref))
+		em.DestroyEntity(ref);
 
     DestroyedEntities.clear();
